% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_diagnosis.R
\name{data_gen_conf_y_analysis}
\alias{data_gen_conf_y_analysis}
\title{Convenience function that will generate one dataset for each configuration
row, print some summaries of the outcome and plot a histogram of the output.}
\usage{
data_gen_conf_y_analysis(data_gen_conf, data_gen_fun)
}
\arguments{
\item{data_gen_conf}{A data.frame containing data generation configurations}

\item{data_gen_fun}{A function for dataset generation using the meadow
interface (TODO)}
}
\value{
Returns nothing but prints text and histograms.
}
\description{
Convenience function that will generate one dataset for each configuration
row, print some summaries of the outcome and plot a histogram of the output.
}
